requirements.txt
livekit>=1.0.6
python-dotenv>=1.1.0
PyAudio>=0.2.14
numpy>=2.0.2